requirements.txt
asttokens==2.2.1
backcall==0.2.0
blinker==1.6.2
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
executing==1.2.0
Flask==2.3.1
Flask-SQLAlchemy==3.0.3
fonttools==4.39.3
greenlet==2.0.2
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipython==8.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
six==1.16.0
SQLAlchemy==2.0.11
stack-data==0.6.2
traitlets==5.9.0
typing-extensions==4.5.0
wcwidth==0.2.6
Werkzeug==2.3.0
zipp==3.15.0
